split everything out to header/impl files for consistency
